Migrated interfaces/types to root folder.
