+++ src/components/debug/debugRouter.js
@@ -8,7 +8,7 @@ class DebugRouter extends Router {
     this.history.listen((location, action)=>{
       console.log(
         `The current URL is ${location.pathname}${location.search}${location.hash}`
-      )
+      );
       console.log(`The last navigation action was ${action}`, JSON.stringify(this.history, null,2));
     });
   }
