+++ src/components/login/login.tsx
@@ -11,23 +11,10 @@ import * as actions from '../../redux/actions/actions';
 //MUI Elements
 import TextField from '@mui/material/TextField';
 import Button from '@mui/material/Button';
+import {UserInfo} from '../../../types';
 
 // @ts-ignore
 import Docketeer from '../../../assets/docketeer-title.png';
-
-interface UserInfo {
-  _id: number;
-  username: string;
-  email: string;
-  phone: string;
-  role: string;
-  role_id: number;
-  contact_pref: string | null;
-  mem_threshold: number;
-  cpu_threshold: number;
-  container_stops: boolean;
-  token: string;
-}
 
 const Login = () => {
   const navigate = useNavigate();
@@ -36,7 +23,37 @@ const Login = () => {
   const updateUser = (userInfo: UserInfo) =>
     dispatch(actions.updateUser(userInfo));
 
-  // callback function invoked when 'login' button is clicked
+   // callback function which will send request to endpoint http://localhost:3000/login and expect
+   const authenticateUser = (username: string, password: string) => {
+    console.log("YOU ARE HERE!")
+    fetch("http://localhost:3000/login", {
+      method: "POST",
+      headers: {
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        username: username,
+        password: password,
+      }),
+    })
+      .then((response) => response.json())
+      .then((data) => {
+        if (Object.prototype.hasOwnProperty.call(data, 'error')) {
+          window.alert(data.error);
+        } else {
+          updateSession(); // loggedIn = true
+          updateUser(data); // update user info in sessions reducer
+          navigate('/');
+        }
+      })
+      .catch((err) => {
+        console.log('Fetch: POST error to /login', err);
+        // created a pop window for wrong username/password
+        window.alert('Wrong Password or Username. Please try Again!');
+      });
+  };
+  
+    // callback function invoked when 'login' button is clicked
   const handleLogin = (e: React.ChangeEvent<HTMLInputElement> | React.ChangeEvent<HTMLFormElement>) => {
     e.preventDefault(); // prevents form submit from reloading page
     console.log("Event:", e)
@@ -54,37 +71,6 @@ const Login = () => {
       console.log("password:", password);
       authenticateUser(username, password);
     
-  };
-
-  // callback function which will send request to endpoint http://localhost:3000/login and expect
-  const authenticateUser = (username: string, password: string) => {
-    console.log("YOU ARE HERE!")
-    fetch("http://localhost:3000/login", {
-      method: "POST",
-      headers: {
-        'Content-Type': 'application/json',
-      },
-      body: JSON.stringify({
-        username: username,
-        password: password,
-      }),
-    })
-      .then((response) => response.json())
-      .then((data) => {
-        if (Object.prototype.hasOwnProperty.call(data, 'error')) {
-          window.alert(data.error);
-        } else {
-          // are the two below functions necessary?
-          updateSession(); // loggedIn = true
-          updateUser(data); // update user info in sessions reducer
-          navigate('/');
-        }
-      })
-      .catch((err) => {
-        console.log('Fetch: POST error to /login', err);
-        // created a pop window for wrong username/password
-        window.alert('Wrong Password or Username. Please try Again!');
-      });
   };
 
   return (
@@ -145,5 +131,5 @@ const Login = () => {
     </div>
   );
 };
-
+}
 export default Login;
