+++ src/components/views/SysAdmin.tsx
@@ -26,7 +26,7 @@ import startNotificationRequester from '../helper/notificationsRequester';
 import initDatabase from '../helper/initDatabase';
 
 // Types and Interface
-import { ContainerObj, StoppedContainerObj, ImageObj, UserObj, VolumeObj, NetworkObj, StateType  } from "./viewsTypes";
+import { ContainerObj, StoppedContainerObj, ImageObj, UserObj, VolumeObj, NetworkObj, StateType  } from '../../../types';
 
 // Container component that has all redux logic along with react router
 const SysAdmin = () => {
