+++ src/redux/reducers/containerListReducer.ts
@@ -1,18 +1,9 @@
 /* eslint-disable no-case-declarations */
 import { PayloadAction } from '@reduxjs/toolkit';
 import * as types from '../constants/actionTypes';
+import { containerState } from '../../../types';
 
-interface stateType {
-  runningList: any[],
-  stoppedList: any[]
-}
-
-const initialState: stateType = {
-  runningList: [],
-  stoppedList: []
-};
-
-export default function (state = initialState, action: PayloadAction<any>) {
+export default function (state = containerState, action: PayloadAction<any>) {
   switch (action.type) {
     case types.ADD_RUNNING_CONTAINERS:
       const newRunningList: object[] = state.runningList.slice();
