+++ src/redux/reducers/notificationReducer.ts
@@ -1,14 +1,8 @@
 import * as types from '../constants/actionTypes';
 import { PayloadAction } from '@reduxjs/toolkit';
+import { notificationState } from '../../../types';
 
-const initialState = {
-  phoneNumber: '',
-  memoryNotificationList: new Set(),
-  cpuNotificationList: new Set(),
-  stoppedNotificationList: new Set()
-};
-
-export default function (state = initialState, action: PayloadAction<any>) {
+export default function (state = notificationState, action: PayloadAction<any>) {
   switch (action.type) {
   case types.ADD_PHONE_NUMBER:
     return {
