+++ src/redux/reducers/sessionReducer.ts
@@ -1,22 +1,7 @@
 import * as types from "../constants/actionTypes";
+import { sessionState } from '../../../types';
 
-const initialState = {
-  _id: "",
-  username: "",
-  email: "",
-  phone: "",
-  role: "",
-  role_id: "",
-  contact_pref: "",
-  mem_threshold: "",
-  cpu_threshold: "",
-  container_stops: "",
-  token: "",
-  isLoggedIn: false,
-  userList: [],
-};
-
-export default function (state = initialState, action: any) {
+export default function (state = sessionState, action: any) {
   switch (action.type) {
     // Change isLoggedIn state variable depending on previous value
     case types.UPDATE_SESSION:
@@ -59,7 +44,7 @@ export default function (state = initialState, action: any) {
     // after logging out, remove all user info from session state
     case types.LOGOUT_USER:
       return {
-        ...initialState,
+        ...sessionState,
       };
 
     default:
