+++ src/redux/reducers/userListReducer.ts
@@ -1,15 +1,8 @@
 import * as types from "../constants/actionTypes";
 import { PayloadAction } from "@reduxjs/toolkit";
+import {userState} from '../../../types';
 
-interface stateType {
-  userList: any[];
-}
-
-const initialState: stateType = {
-  userList: [],
-};
-
-export default function (state = initialState, action: PayloadAction<any>) {
+export default function (state = userState, action: PayloadAction<any>) {
   switch (action.type) {
     // Change isLoggedIn state variable depending on previous value
     case types.UPDATE_USER_LIST:
