+++ src/redux/reducers/userReducer.ts
@@ -1,19 +1,7 @@
 import * as types from "../constants/actionTypes";
+import {userReducerState } from '../../../types';
 
-const initialState = {
-  name: "",
-  email: "",
-  phone: "",
-  role: "",
-  role_id: "",
-  contact_pref: "",
-  mem_threshold: "",
-  cpu_threshold: "",
-  container_stops: false,
-  isSysAdmin: false,
-};
-
-export default function (state = initialState, action: any) {
+export default function (state = userReducerState, action: any) {
   switch (action.type) {
     case types.UPDATE_ALL:
       return;
